+++ snapatac2-core/src/utils.rs
@@ -6,7 +6,8 @@ use std::io::{BufWriter, Write};
 use std::str::FromStr;
 use anyhow::{Result, Context};
 
-use bed_utils::bed::{BEDLike, NarrowPeak, merge_bed_with};
+use bed_utils::bed::{BEDLike, NarrowPeak, merge_sorted_bed_with};
+use bed_utils::extsort::ExternalSorterBuilder;
 
 pub fn merge_peaks<I>(peaks: I, half_window_size: u64) -> impl Iterator<Item = Vec<NarrowPeak>>
 where
@@ -24,17 +25,19 @@ where
         result
     }
 
-    merge_bed_with(
-        peaks.map(move |mut x| {
-            let summit = x.start() + x.peak;
-            x.start = summit.saturating_sub(half_window_size);
-            x.end = summit + half_window_size + 1;
-            x.peak = summit - x.start;
-            x
-        }),
-        iterative_merge,
-        None::<&str>,
-    )
+    let input = peaks.map(move |mut x| {
+        let summit = x.start() + x.peak;
+        x.start = summit.saturating_sub(half_window_size);
+        x.end = summit + half_window_size + 1;
+        x.peak = summit - x.start;
+        std::io::Result::Ok(x)
+    });
+    let input = ExternalSorterBuilder::new()
+        .with_compression(2)
+        .build().unwrap()
+        .sort_by(input, BEDLike::compare).unwrap()
+        .map(|x| x.unwrap());
+    merge_sorted_bed_with(input, iterative_merge)
 }
 
 pub fn clip_peak(mut peak: NarrowPeak, chrom_sizes: &crate::preprocessing::count_data::ChromSizes) -> NarrowPeak {